Add files via upload
The title of this assignment was "Evaluate Techniques for Wifi Locationing".    For this task, we were asked  investigate the feasibility of using "wifi fingerprinting" to determine a person's location in indoor spaces.  We were provided with a large dataset of wifi fingerprints for a multi-building industrial campus with a location (building, floor, and location ID) associated with each fingerprint.  Once this data was uploaded into RStudio, we had to develop multiple machine learning models and evaluate their application to the problem of indoor locationing via wifi fingerprinting.