Add files via upload
Two small matrices with sentiment toward the iPhone and Samsung Galaxy smartphones were provided to the cohort.  From this information, the goal was to build machine learning models that understood the patterns in the two smaller matrices.  These models would then be used with the large matrix file downloaded from Amazon Web Services to predict sentiment for the two smartphones..